Add support for Homeowner Keypads
Lutron allows keypads to be added to rooms which are not
tied to physical devices, and can be controlled soley through
the LutronConnect App.  These are represented in the
`DbXmlInfo.xml` as a `<Device />` with the attribute
`DeviceType="HOMEOWNER_KEYPAD"`.

This commit adds support for those keypads during the
area parsing, and ups the version number of pypi.