Fix handling of connection to Main Repeater going away
Previously we'd die in many places, or get stuck in some blocking calls
in telnetlib.

This is probably still not great, but in killing the MR several times,
we recover ok.

TODO: Still missing is to check if the xml info changed and reload that.
It will be harder since our clients could have registered callbacks, etc.