Change the legacy UUID generation to use a separate property named legacy_uuid rather than hiding it behind the uuid property.
This would allow users to know when UUID is actually set, and make informed decisions about if/when to use a legacy synthesized UUID rather than assume this is something that is coming from the XML.  The second a user updates their firmware to a later version, uuid() will start returning different values.  legacy_uuid should be more stable in that respect.

Note: Integration ids can/do change when significant changes are made to the programming of a system which is why UUIDs exist in the first place.