Generate Legacy UUIDs using IntegrationIDs and other information. (#90)
* Add support for generating UUIDs with older firmware.

Much older Lutron firmware doesn't have UUIDs.  We generate legacy UUIDs using the integration IDs and component numbers.  These are accessible via the legacy_uuid property and should work for all versions of firmware.

Note: Integration ids can/do change when significant changes are made to the programming of a system which is why UUIDs exist in the first place.